fixes and changes for msvc toolchains.
fixed compiler warnings.
properly exported functions for dlls.
replaced a lot of posix code.

+++ include/moshi/json.h
@@ -127,7 +127,7 @@ int json_maybe_get_int64_array(const_str_t & json, int offset, T on_array) {
             } else {
                 skipping = true;
             }
-            offset = (end - start) + offset;
+            offset = (int) (end - start) + offset;
         } else {
             skipping = true;
